+++ scripts/INSTALL_prerequisites_with_apt-get.sh
@@ -84,10 +84,10 @@ echo "Installing python APT packages"
 # we will use pip for most
 # some extra package needed for jupyter
 qt=pyqt5
-${APT_GET_INSTALL} python-dev python-pip python-tk python-${qt} python-${qt}.qtsvg python-${qt}.qtwebkit
+${APT_GET_INSTALL} python3-dev python3-pip python3-tk python3-${qt} python3-${qt}.qtsvg python3-${qt}.qtwebkit
 
 echo "Installing CIL pre-requisites..."
-${APT_GET_INSTALL} cython python-h5py python-wget
+${APT_GET_INSTALL} cython3 python3-h5py python3-wget
 # dependency for ASTRA-toolbox autotools-dev automake autogen autoconf libtool
 
 echo "Run INSTALL_python_packages.sh after this."